chore: move or delete DTO files
